pyproject.toml
[tool.poetry]
name = "muser"
version = "0.1.0"
description = "Instrument trading/selling site"
authors = ["Zuiluj <julzllanillo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2.3"
requests = "^2.22"
djangorestframework = "3.9.4" 
django-rest-auth = "^0.9.5"
mongoengine = "^0.18.2"
django-rest-framework-mongoengine = "^3.3"
django-mongoengine = "^0.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
